+++ src/bin_node/node_shared_arg.ml
@@ -199,10 +199,11 @@ module Term = struct
       "The directory where the Tezos node will store all its data. Parent \
        directories are created if necessary."
     in
+    let env = Arg.env_var Node_config_file.data_dir_env_name in
     Arg.(
       value
       & opt (some string) None
-      & info ~docs ~doc ~docv:"DIR" ["data-dir"; "d"])
+      & info ~docs ~env ~doc ~docv:"DIR" ["data-dir"; "d"])
 
   let config_file =
     let doc = "The main configuration file." in